+++ v0.4/dev/prod/package.json
@@ -33,9 +33,8 @@
     "svgo-loader": "~3.0.0"
   },
   "dependencies": {
-    "@anticrm/plugin-ui": "~0.2.0",
     "@anticrm/platform": "~0.2.0",
-    "@anticrm/plugin-ui-impl": "~0.2.0",
+    "@anticrm/ui": "~0.2.0",
     "svelte": "~3.37.0",
     "@anticrm/plugin-login": "~0.2.0",
     "@anticrm/plugin-login-impl": "~0.2.0"
