+++ v0.4/platform/ui/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@anticrm/plugin-ui-impl",
+  "name": "@anticrm/ui",
   "version": "0.2.0",
   "main": "src/index.ts",
   "author": "Anticrm Platform Contributors",
@@ -32,11 +32,10 @@
   },
   "dependencies": {
     "@anticrm/platform": "~0.2.0",
-    "@anticrm/plugin-ui": "~0.2.0",
     "fast-equals": "~2.0.1",
     "@anticrm/status": "~0.2.0",
     "@anticrm/sparkling-theme": "~0.2.0",
     "svelte": "~3.37.0",
-    "@anticrm/sparkling-components": "~0.2.0"
+    "@anticrm/sparkling-controls": "~0.2.0"
   }
 }
