+++ v0.4/platform/ui/src/components/internal/ErrorBoundary.js
