+++ v0.4/platform/ui/src/routes.ts
@@ -13,12 +13,14 @@
 // limitations under the License.
 //
 
-import { ApplicationRouter, Location } from '@anticrm/plugin-ui'
-import { parseHash, parsePath, parseQuery } from './location'
+import { ApplicationRouter, Location } from './types'
+import { parseHash, parsePath, parseQuery, locationToUrl, parseLocation } from './location'
 
 import { deepEqual } from 'fast-equals'
+import { getContext, onDestroy, setContext } from 'svelte'
+import { writable, Readable, derived } from 'svelte/store'
 
-export class Router<T> implements ApplicationRouter<T> {
+class Router<T> implements ApplicationRouter<T> {
   private readonly pattern: string
   private segments: string[] = []
   private queryNames: string[] = []
@@ -357,4 +359,80 @@ export class Router<T> implements ApplicationRouter<T> {
   clearChildRouter (): void {
     this.childRouter = undefined
   }
+}
+
+function windowLocation (): Location {
+  return parseLocation(window.location)
+}
+
+const locationWritable = writable(windowLocation())
+window.addEventListener('popstate', () => {
+  locationWritable.set(windowLocation())
+})
+
+const location: Readable<Location> = derived(locationWritable, (loc) => loc)
+
+function subscribeLocation (listener: (location: Location) => void, destroyFactory: (op: () => void) => void): void {
+  const unsubscribe = location.subscribe((location) => {
+    listener(location)
+  })
+  destroyFactory(unsubscribe)
+}
+
+function navigate (newUrl: string): void {
+  const curUrl = locationToUrl(windowLocation())
+  if (curUrl === newUrl) {
+    return
+  }
+  history.pushState(null, '', newUrl)
+  locationWritable.set(windowLocation())
+}
+
+function navigateJoin (
+  path: string[] | undefined,
+  query: Record<string, string> | undefined,
+  fragment: string | undefined
+): void {
+  const newLocation = windowLocation()
+  if (path != null) {
+    newLocation.path = path
+  }
+  if (query != null) {
+    // For query we do replace
+    const currentQuery = newLocation.query || {}
+    for (const kv of Object.entries(query)) {
+      currentQuery[kv[0]] = kv[1]
+    }
+  }
+  if (fragment) {
+    newLocation.fragment = fragment
+  }
+  navigate(locationToUrl(newLocation))
+}
+
+const CONTEXT_ROUTE_VALUE = 'routes.context'
+
+export function newRouter<T> (
+  pattern: string,
+  matcher: (match: T) => void,
+  defaults: T | undefined = undefined
+): ApplicationRouter<T> {
+  const r: Router<any> = getContext(CONTEXT_ROUTE_VALUE)
+  const navigateOp = (loc: Location): void => {
+    navigate(locationToUrl(loc))
+  }
+  const result = r ? r.newRouter<T>(pattern, defaults) : new Router<T>(pattern, r, defaults, navigateOp)
+  result.subscribe(matcher)
+  if (!r) {
+    // No parent, we need to subscribe for location changes.
+    subscribeLocation((loc) => {
+      result.update(loc)
+    }, onDestroy)
+  }
+  if (r) {
+    // We need to remove child router from parent, if component is destroyed
+    onDestroy(() => r.clearChildRouter())
+  }
+  setContext(CONTEXT_ROUTE_VALUE, result)
+  return result
 }
