+++ v0.4/plugins/login-impl/package.json
@@ -25,9 +25,8 @@
     "@anticrm/platform": "~0.2.0",
     "svelte": "~3.37.0",
     "@anticrm/status": "~0.2.0",
-    "@anticrm/plugin-ui": "~0.2.0",
+    "@anticrm/ui": "~0.2.0",
     "@anticrm/plugin-login": "~0.2.0",
-    "@fingerprintjs/fingerprintjs": "~3.1.0",
     "@anticrm/rpc": "~0.2.0",
     "@anticrm/sparkling-controls": "~0.2.0"
   }
