+++ v0.4/plugins/login/package.json
@@ -20,7 +20,7 @@
   },
   "dependencies": {
     "@anticrm/platform": "~0.2.0",
-    "@anticrm/plugin-ui": "~0.2.0",
+    "@anticrm/ui": "~0.2.0",
     "@anticrm/status": "~0.2.0"
   }
 }
